ipa: libipa: lsc: Quantize gains in IPA
The libIPA Lsc algorithm currently assumes the register format of
the RkISP1 platform (16 bits UQ2.10 format), as that's where the
implementation has been derived from.

Make lsc::Components use floats as the default exchange type between the
LscAlgorithm and the IPA module and move the gains quantization to the
plaltform IPA module.

+++ src/ipa/libipa/lsc_base.h
@@ -26,7 +26,7 @@ namespace ipa {
 
 namespace lsc {
 
-using Components = std::map<std::string, std::vector<uint16_t>, std::less<>>;
+using Components = std::map<std::string, std::vector<float>, std::less<>>;
 using ComponentsMap = std::map<unsigned int, Components>;
 
 } /* namespace lsc */
